requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
docopt==0.6.2
Flask==1.1.2
fsspec==0.8.7
future==0.18.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.36.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
Pillow==8.1.1
pipreqs==0.4.10
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytorch-lightning==1.2.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp38-cp38-linux_x86_64.whl
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
yarg==0.1.9
yarl==1.6.3

